requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==42.0.5
python-dotenv==1.0.1
Werkzeug==3.0.1
PyJWT==2.8.0
google-generativeai==0.4.1